pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "tastytrade/__init__.py"

[project]
name = "tastytrade"
description = "An unofficial, typed, asynchronous Python SDK for Tastytrade!"
readme = "README.md"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Framework :: AsyncIO",
    "Framework :: AnyIO",
    "Framework :: Trio",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS :: MacOS X",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: Unix",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP",
    "Typing :: Typed",
]
requires-python = ">=3.11"
license = {file = "LICENSE"}
authors = [
  {name = "Graeme Holliday", email = "graeme@tastyware.dev"},
]
dependencies = [
    "anyio>=4.12.1",
    "httpx>=0.28.1",
    "httpx-ws>=0.7.2",
    "pandas-market-calendars>=5.1.1",
    "pydantic>=2.11.9",
]
dynamic = ["version"]

[project.urls]
Homepage = "https://github.com/tastyware/tastytrade"
Documentation = "https://tastyworks-api.rtfd.io"
Funding = "https://github.com/sponsors/tastyware"
Source = "https://github.com/tastyware/tastytrade"
Changelog = "https://github.com/tastyware/tastytrade/releases"

[dependency-groups]
dev = [
    "pytest>=8.3.3",
    "pytest-cov>=5.0.0",
    "ruff>=0.13.2",
    "pyright>=1.1.401",
    "sphinx>=7.4.7",
    "enum-tools[sphinx]>=0.12.0",
    "autodoc-pydantic>=2.2.0",
    "proxy-py>=2.4.9",
    "sphinx-immaterial>=0.12.5",
    "mypy>=1.18.2",
    "pandas-stubs>=2.2.2.240807",
    "trio>=0.32.0",
]

[tool.setuptools.package-data]
"tastytrade" = ["py.typed"]

[tool.setuptools.packages.find]
where = ["tastytrade"]

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pyright]
pythonVersion = "3.11"
include = ["tastytrade", "tests"]
exclude = ["**/__pycache__"]
typeCheckingMode = "strict"
reportPrivateUsage = false

[tool.mypy]
strict = true
warn_unused_ignores = false
